Changes for Flutter App

+++ voyager-datasync/src/main/java/org/voyager/sync/model/flightradar/search/FlightTimeFR.java
@@ -7,5 +7,5 @@ public class FlightTimeFR {
     String aircraft;
     String time;
     Long timestamp;
-    Long offset;
+    Integer offset;
 }
